fixed configuration issues; tidied up scripts storage
